feat: Initial data switched from docker to application startup

+++ src/main/java/kornas/moviecharactersapi/Repositories/FranchiseRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FranchiseRepository extends JpaRepository<Franchise, Integer> {
+public interface FranchiseRepository extends JpaRepository<Franchise, Long> {
 
 }
